Use Db for task management
This change adjusts the Tasks type to work with a Db of Task objects
instead of holding a vector of them itself. As such, this type is no
longer concerned with implementing low-level task management, but only
with providing additional functionality on top of it, such as the higher
level undo and redo operations.