Keep track of task IDs in Db
So far our Db has merely piggy-backed on the assumed-to-be-unique task
IDs that we have used so far. As a first step towards taking over ID
management, this change introduces some ID management functionality to
the Db type. Specifically, we maintain a set of IDs that we keep in sync
as objects are added to the database and removed from it.