Add support for allocating and freeing task IDs to Db
We want our database to eventually take over ID management completely,
allowing us to eliminate the tight coupling between tasks and their
(globally unique) IDs.
To bring us one step closer to this goal, this change introduces the
means for reserving, allocating, and freeing of IDs to the Db type.