Store items along with IDs in database
So far we relied solely on the Idable trait for everything pertaining
database IDs. This change adjusts the database to store IDs along with
items. IDs are still retrieved from the item (by virtue of said trait),
but we are making a first step towards having the database take over all
tasks pertaining ID management. Along with this more or less internal
change, we also adjust public methods on the Db type to expose the item
ID, where applicable. Iteration over the database's items, for example,
now yields (id, item) tuples.
Because IDs are meant to be read-only, we remove the IndexMut and Index
implementations. In their stead we introduce the new Db::get (and
get_mut) API in order to easily provide access to an item's ID while
preventing mutation of it. Because it is fallible, we clean up the
earlier behavior where we could conceivably have panics all over the
place, provided an invalid index.