Expose task IDs through our View
We have the concept of a View, which effectively provides a view of
tasks matching certain criteria. So far, iterating over such a view only
yielded tasks themselves. That was fine, because tasks had IDs directly
associated with them.
Given that this association will be removed in the future, this change
adjusts the iterator to now yield tasks alongside their IDs instead,
making the same information available to clients.