Make TaskListBox::selected_task return a task reference
The TaskListBox::selected_task returns a copy of the selected task. That
is arguably a bit wasteful, as some call-sites only need temporary
access to the task.
Because cloning can always happen later on, this adjusts the method to
return a reference to the selected task instead, foregoing any deep
copies.