Adjust dialog data to store task ID
This change adjusts the dialog's internally used Data type to store the
task ID in addition to the task itself. This step comes in preparation
for removing the task ID member from Task objects themselves and ensure
that we have access to the task ID once the removal has occurred.