Remove Task::id() method
With recent changes we have eliminated all but a few clients of the
Task's id method. This change replaces the remaining few and removes the
method from the type. In so doing the Task::id attribute is now truly
internal to the containing module. Upcoming changes will continue the
work and remove the attribute from there as well.