Make Db::{insert,push} return the Id of the added item
When we insert a new element into our database, it would be nice to
receive some kind of handle to use to access it in the future. That is,
of course, what the database's IDs are meant for to begin with.
With this change we adjust our Db::insert and Db::push methods to return
the ID of the inserted element to the caller.