Adjust Db::remove method to return (id, item) tuple
So far the Db::remove method has only returned the item that was
removed. Soon items will no longer implement the Idable trait and,
hence, may loose the ability to identify themselves. As such, it makes
sense for the method to return the item along with its ID, should
callers need it. This change accomplishes that.