Make TaskOp::Add work with optional ID
The operation for adding a task currently embeds the ID of the task.
That works while each task carries its own ID, but once that changes we
may not have an ID available until the task actually got inserted into
the database.
To prepare for the case of tasks no longer carrying IDs with them, this
change adjusts the TaskOp enum's Add variant to store an optional Id
instead.