Adjust add_task function to accept optional Id
We use the add_task function both when adding a new task but also when
re-adding a task after redoing an operation. In the latter case, we may
already have an ID that we want to use from a previous removal. To make
it possible to preserve this ID, this change adjusts the add_task
function to accept an optional Id, which it then passes on to
Tasks::insert or Tasks::push, respectively, which were adjusted with
such an argument earlier.