Use stashed ID in TaskOp::undo
The cases of task addition and task update when triggered from the
operation undo path are still relying on the task's ID. By now, however,
we do have the associated database ID available as well. With this
change we switch these code paths over to using the latter, in
preparation for the eventual removal of the identifier directly
associated with the task.