Remove Idable constraint from various Db methods
So far our database was still relying on its items implementing the
Idable trait for one reason or another. By now, however, we have all the
pieces in place to get rid of this dependency. To that end, this change
removes the Idable constraint from all the Db type's methods, with the
exception of one of the Db::try_from_iter constructor.