Introduce Db::from_iter constructor
We can instantiate Db objects with the try_from_iter constructor. This
fallible constructor expects the caller to provide unique IDs (currently
implicitly by virtue of items having to implement the Idable trait). For
testing purposes that is unnecessarily complicated. In fact, there most
of the time we do not want to have the test come up with those IDs at
all.
This change introduces a new construtor, Db::from_iter, that tests may
use to instantiate a Db from a list of items and have it take care of ID
assignment.