Adjust Db::try_from_iter to accept IDs to use
This change adjusts the Db::try_from_iter constructor to accept an
iterator of (id, item) tuples instead of one working with items that
implement the Idable trait. In so doing we get ready for the removal of
the Idable trait altogether and allow clients to have a say about IDs to
use in the database (which will eventually become a requirement when we
persist IDs).