Remove Task::id member
Earlier changes have removed most references to the Task's id member.
This change takes the next logical step and removes the member
altogether, replacing the few remaining references to it with database
ID instances instead. In addition to the member variable, the Idable
trait is removed as well, as it was only ever implemented to Task
objects and that is no longer possible.