Make Id::from_unique_id constructor private again
At some point in the past we had to make the Id::from_unique_id
constructor available to various parts of the program in order to
perform a step-wise conversion away from using Task::id, which required
the direct construction of database IDs. By now we have removed the
Task::id member and nobody except for the Db type itself has to create
database IDs anymore. Hence, with this change we make said constructor
private again.