Introduce TasksMeta type
This change introduces a new serializable struct, TasksMeta. This type
is comprised of meta data accompanying tasks themselves that will have
more significance with upcoming changes. Among other things it can be
used to persist (and restore) tasks in their original order when the
task storage itself does not preserve this information.
Given that objects of this type represent task state, we add one to the
ser::state::TaskState type. Note that right now this member is ignored
during serialization while we swap out the backend to use for task
storage.