Add support for saving tasks in a directory
This change introduces the means to save tasks in a different "format":
instead of dumping out a large database containing all tasks (along with
task related information) into a single file, we instead store each task
in a different file, with the file name determined by the task's ID.

This format will have multiple advantages. For one, it will enable us to
actually only persist changes to tasks (as opposed to saving all tasks
unconditionally, as we currently do). Furthermore, this format is
conceptually much closer to the vdir storage format, which we could
conceivably use for synchronizing tasks with a remote server. Lastly,
because we identify tasks by ID in a single directory, we can rely on
the file system to make sure that task IDs stay unique -- an invariant
that cannot be enforced when just stashing away IDs in a big JSON blob.
This latter property is believed to ultimately help us persist task IDs
to begin with (as currently they are solely an in-memory construct).
That in turn would allow us to persist our undo stack as well.