Plumb through tasks root directory
In order to be able to save tasks as files below a directory, we need to
plumb the path to said directory through to the layer that persists task
state.
This change takes care of that. We treat the path to this directory
similarly to how we treat the UI configuration file path, for example,
and pass it around in the same way.