Save tasks in directory
This change finally hooks up the task saving logic to store tasks in
individual files below the inferred task directory. Note that currently
this is just a second way in which tasks are saved (in addition to the
task configuration file) and write-only as such, until we fully convert
over to also loading tasks from this directory.