Add support for loading tasks from a directory
This change adds support for loading tasks from a directory to the
program. That means we now have the means to load what we stored earlier
now. Please note, however, that this functionality is not yet wired up
properly: we still load the tasks we work with from a single JSON file.