Stop saving tasks into a file
Now that we are loading tasks from a directory as opposed to a file, we
no longer need to carry around the functionality for saving them into a
file to begin with. So far we have done both, saved to a file and to a
directory.
This change takes the logical step and removes support for saving them
into a file.