Remove ser::state::TaskState::templates member
This change removes the templates member of the ser::state::TaskState
type. This member had a purpose while we were savings the task state
into a file, but in the new world we save everything into a directory
and there we rely on the task meta data for the same information.