Remove no longer needed task configuration file logic
A bunch of the main program but also test code still has logic for
working with the task configuration file -- the file the formerly was
used for saving tasks to and loading them from. Given that we switch
over to saving tasks in a directory now, this logic is superfluous.
This change removes it.