Store "toggle tag" in UI state
This change introduces the notion of the "toggle tag" to the UI state.
This attribute, as the name tries to convey, represents a tag that the
user will be able to toggle through the UI. Currently the user can
toggle the completion tag and that logic is hard wired and not
configurable. This new tag will replace the existing logic with a user
configurable one. Subsequent changes will plumb it through the remaining
layers.