Store "toggle tag" in TaskListBox
The TaskListBox is the entity that currently reacts to key presses that
toggle the completion state of a task. As such, it will also be the
entity that takes care of honoring the user-configurable "toggle tag"
once it replaces the completion tag.
This change makes sure to store said tag inside objects of the type.
Plumbing is added to pass the tag instance through the various layers.