Store Task::tags in set instead of map
So far we have used a map of (TagId, Tag) tuples for keeping track of
the tags associated with a task. However, the TagId part of the tuple is
really only required to facilitate comparisons.
Given that we aim to make such ID more of a lower-level detail, this
change switches to us using a set of Tag objects directly instead. To
make this work we ensure that equality checks on such objects work as
expected and that they can be ordered properly (as we are currently
using a BTreeSet).