Make Task::summary a private attribute
Our Task type has one public attribute: the summary. Other than
laziness, there is no really good reason for that fact and it is not
nice from an encapsulation point of view.
To fix this problem, this change makes said attribute private and
replaces accesses with calls to the newly added getter and setter
methods.