Stop using serde_json::from_reader for deserialization
With upcoming changes we intend to make our state loading and saving
functionality async. Currently, the deserialization aspect of it relies
on the serde_json::from_reader function. Unfortunately, there exists no
async counter part to this function and so we have to fall back to
reading the entire file into memory and deserializing directly from the
buffer.
With this change we switch the deserialization code over from using
from_reader to from_slice.