Make run_prog() function async
So far all of our setup code was running outside an async runtime and
only the main loop was async. With this change we declare run_prog() as
async, moving async runtime creation before it. This step is necessary
in order to make it possible for us to make all of our state loading and
saving async, some of which happens on the setup path.