Declare state loading and saving logic as async
So far all our state loading and saving functionality was synchronous,
using potentially long blocking system calls. That is arguably not
great, because at least the saving of state can actually be triggered
from an async context, where the runtime will be blocked and prevented
from executing other async functionality while a blocking system call is
made.
With this change we prepare for fixing this issue. Specifically, we
change all functions involved in loading and saving of state along with
all callers to be async functions. Note that are the lowest level, we
are still using blocking file system primitives. This will be changed
subsequently.