Adjust load_state_from_file to return an Option
The load_state_from_file function creates a default representation of
the state object to load if the file to load it from is not actually
present. This behavior makes the function unnecessarily hard to use with
upcoming changes, that want to cater to the "not found" case
differently, while expecting the same behavior under all other
conditions.
In an attempt to make this function more universally usable, this
adjusts the return value to contain an Option instead, which is simply
set to None when no file is present. Callers can easily map that to the
default representation if they so please.