Only save state if it actually changed
With the switch from storing all tasks in a single large file to using
one file per task, it makes much more sense for us to be cleverer about
when to actually write data into such a file. Specifically, given the
potential large amount of tasks with only few ever changing between save
operations, it is rather wasteful to write all the tasks out all the
time.
There are probably many ways to introducing smarts into the system in an
attempt to reduce the number of writes. With this change we do the
simplest but arguably most robust and effective thing: we read the
content of the file and deserialize it and only if the state to write
out is different from this already-present state do we actually persist
the new data.
An alternative approach could be to track whether a task (or any other
to-be-persisted object) changed since it was loaded. However, while
perhaps simple sounding, implementing such logic correctly is a complex
effort. For example, we have to take into consideration the file we
loaded from and whether we are saving to the same one. There are also
interesting interactions with our undo and redo system possible. Given
these complications, this avenue was ultimately discarded in favor of a
live comparison against the on-disk contents.