Introduce Template::with_serde constructor
Currently the Template type sports a From implementation for the
conversion from the corresponding ser type. That is counter to virtually
all other types, which instead have a constructor called with_serde.
That is unintuitive and inconsistent and as such, this change introduces
such a constructor and makes it replace the earlier From impl.