Make Template::new constructor private
This change adjusts the Template::new constructor's visibility from
public to private. It was really only used for testing purposes, and for
those cases it is now possible to use a (testing-only) Extend impl on
the Templates type, which works with strings instead of fully blow
templates, and which will both take care of creating templates and then
inserting them.