Rename id module to uid
This change renames the id module to uid. We want to introduce a new
module named id with upcoming changes, and so we need this name to be
available. The existing module is a thin wrapper around the Id type as
provided by the uid crate, which is what the new name reflects.