Introduce and use AllocId trait
The Db has basically hand rolled a custom allocator for Id objects based
on a set. In the future we want to use such an allocator not only for
database IDs, but also for other ones. As such, it makes sense to
decouple the logic from the Db type and make it reusable.
This change does that, by introducing a new trait, AllocId, which
provides the corresponding allocation, reservation, and release methods
and which is implemented for the BTreeSet type.