Add Entry associated type to AllocId trait
With upcoming changes we would like to implement the AllocId trait not
only for a BTreeSet but also a BTreeMap. However, a map imposes
different requirements on the implementor than a set does. For example,
when using the same of methods on it, insert specifically, we run into a
chicken-egg problem when allocating an ID, at least if the value type is
meant to contain the ID as well: insertion requires a fully instantiated
value, but value instantiation requires an ID. Said ID, however, would
only be available after insertion.
Luckily for us, Rust's BTreeMap provides the Entry API, which allows us
to solve this conundrum nicely, by returning an entry object reference
from the allocation path along with the allocated ID; clients can then
construct the value with the ID and use the entry object to insert it
into the map.
To make all of this feasible, we need to amend our AllocId trait a bit.
Specifically, we introduce a new associate type, Entry, an object of
which is being returned on allocation and reservation. This change
introduces this type and adjusts the API surface accordingly to use it.