Implement AllocId for BTreeMap
With upcoming changes we would like to have ID allocation not only on a
set but also on a map type. In our case would be possible to have two
sets instead, one solely dedicated to allocation and another holding the
actual values of interest, but that makes it possible for the two
objects to get out of sync: if a value is removed but not freed or vice
versa, for example.
The usage of a map effectively prevents that from happening, because if
a value is removed from it, so is its ID. To that end, this change
implements the AllocId trait for the BTreeMap type.