Store templates in a map instead of a set
We currently store all of our templates in a set. That has served us
well so far. However, with upcoming changes we would like to switch from
globally-unique (read: program-wide) IDs to those unique only in a
certain "space" (the Templates type, in our case).
One requirement for this switch is that we have the means to allocate
IDs in said space. Earlier changes have made it possible for us to
allocate IDs stored in a map, where the keys are IDs and the values can
be user provided.
In order to piggy-back on this work, this change adjusts the Templates
type to store all its templates in a map of <id, template> pairs instead
of a set of templates. For now we still use globally-unique IDs, but
subsequent changes will make the switch to actually allocating them
instead.