Replace specialized task ID serialization with more general one
The only ID type that so far has required serialization is that for task
IDs. As such, we only implemented the ToSerde trait for those IDs. With
upcoming changes we would like to serialize other IDs as well.
To make this possible, this change replaces the specialized task ID
ToSerde impl with a more general one applicable to all Id objects.