Use id::Id type in templates
This change adjusts the Template struct to use the id::Id type for its
IDs, instead of the uid::Uid type used so far. The latter is used for
globally-unique IDs, which impose many requirements that make our code
unnecessarily complicated. Objects of the id::Id, on the other hand, are
unique only within some defined "space". In our case, we ensure
uniqueness within the purview of one Templates instance.