Introduce Templates::instantiate_serde()
This change introduces a new method on the Templates type to instantiate
a tag: instantiate_serde(). Contrary to other instantiate* methods, this
one accepts a serialized IDs. As such, it needs to be fallible. A first
set of clients is adjusted to use this method over the existing
instantiate(), which works on a translated ID instead.