Adjust View::iter() to perform internal task iteration
Currently we use external iteration in a bunch of locations, such as in
order to traverse all tasks represented by a view. That is mostly
because it is the default model that Rust employs and so it has the
least friction.
This change marks the first reconsidering that choice. Specifically, it
converts the View::iter() method over to using (a variant of) internal
iteration, where the caller passes in a callback that is invoked on an
iterator over the view's tasks, but without said iterator ever leaving
the calling context.
We go down this route, because we want to switch from using the
RefCell<Tasks> pattern to making RefCell internal to the Tasks type.
There is not really a way to hand out iterators, that is, objects
referencing the object wrapped by a RefCell, and return them to a caller
while having a borrow that is only valid while the called method is
active. Using internal iteration as described solves this use case.