Switch to using std::cell instead of cell crate
With View::iter() converted to using internal iteration, the one user of
the cell crate dependency is gone. As such, this change replaces all
usages of cell::* with their counterparts in std::cell::* and then
removes the dependency altogether. The crate has a subtle unsoundness
and as such usage is discouraged, so this is a welcome removal.