Make Tasks::iter() use internal iteration
Similar to what we did earlier with View::iter(), this change adjusts
the Tasks::iter() method to use a variant of internal iteration, where
the caller provides a closure that operates on an iterator over all
tasks, as opposed to said iterator being returned to the caller
directly. In so doing we prepare the Tasks type to use an internal
RefCell, as opposed to the external one (RefCell<Tasks>) currently
employed.