Use RefCell inside Tasks type
This change adjusts the Tasks type by making it contain a RefCell of a
newly introduced TasksInner struct. In so doing we set the stage for
removing the RefCell<Tasks> pattern from the code base. This pattern is
problematic, because it can easily lead to borrowing conflicts that
panic the program. By having the RefCell hidden behind a public API
surface, we can mitigate such potential problems much more easily,
because violations crossing API boundaries now have to be very
intentional.