Use internal Task tag iteration
Similar to what we did for Tasks::iter() and View::iter() with earlier
patches, this change adjusts Task::tags() to use a form of internal
iteration. We rely on exactly the same pattern these other iteration
methods use, where the function accepts a closure that is passed an
iterator that never leaves the calling context.
This change comes in preparation for a larger refactoring which makes us
work with reference counted Task objects, instead of the current
approach which heavily relies on IDs for "sharing".