Make Task::summary() return String
The Task::summary() method currently returns a string reference to the
caller. Similar to the Tasks type, we need Task to have internal
mutability by using a RefCell. Once we do that, we can no longer simply
hand out a reference to callers.
This change adjusts the Task::summary() method to return a fully blown
String object instead. A lot of the call sites already create an owned
object from the string, and just in general we are should not be overly
sensitive to a few more allocations happening to be worried about this
clone operation.