Make Task::templates() return Rc<Templates>
For better or worse, each Task object carries a reference to the
Templates instance from which its tags were (meant to be) instantiated.
Currently, it provides access to a reference to it via the
Task::templates() method.
Returning a reference like this will become problematic once we wrap
Task internals in a RefCell and so with this change we return a fully
blown Rc<Templates> object instead.