Make Task internally mutable
This change makes the Task type internally mutable, by making it a
RefCell of the internal TaskInner. In so doing we pave the way for
sharing of Tasks object directly, as opposed to relying on a special ID
(and an associated lookup) for that purpose. Internal mutability is a
prerequisite for sharing, because we already modify tasks from a bunch
of places and so read-only access is insufficient.