Store Task objects in Rc
We use task IDs for "identification" purposes but also as a poor man's
reference: they are trivially shareable and enable lookup of the
referenced task.
However, after some deliberation we have come to the conclusion that
this is not the best approach: there is no guarantee that a task ID is
valid and that necessitates the existence of a bunch of error cases. At
a conceptual level, there is also nothing preventing a "use after free"
kind of scenario, where a task is deleted and a new task with the same
ID is added. Given that we do not prevent task ID reuse currently, this
could lead to a wrong task being referenced. Lastly, there is the fact
that looking up a task by its ID is just conceptually an unnecessary
indirection: why not reference the Task object directly?
This change marks a first step (well, depending on where we start
counting...) towards shareable Task objects replacing task IDs.
Specifically, we start working with Rc<Task> objects, which can easily
be cloned. We treat each Rc pointing to the same heap allocation as
having the same identity (i.e., "being" the same object).
Note that for the time being we still have task IDs floating around. The
transition away from task ID usage will take a few changes, owing to
their pervasiveness in the code base.