Add support for configurable equality checks to Db type
Our Db type supports lookup of objects stored in it (currently only of
Task type) by ID. As motivated earlier, we want to get rid of the
concept of an ID for reference purposes. With the recent switch to using
Rc<Task>, what we actually want to use for lookup is the address of the
heap allocated Task object, as it constitutes the object's "identity".

To make that happen we need a way to tell the Db how to compare items.
To that end, this change introduces another generic parameter to it. It
is assumed to reference a type providing a method that can be used to
compare two items for equality. By default, this type just forwards to
the store item's Eq implementation.
Along with that parameter, we introduce the new Db::find_item() method,
which honors this input to support direct item lookup (as opposed to one
through an ID).