Introduce Cmp impl for comparing Rc<Task>
A recent change to the Db type made it accept an additional generic
parameter that customizes the logic for finding items when using the
Db::find_item method.
This change now provides an type implementing the corresponding trait,
Cmp, to the Db instance used within the Tasks object. This type, CmpRc,
can be used for checking Rc pointers for equality, instead of comparing
the content they refer to.