Simplify tasks::TaskOp::Update variant
This change simplifies the Update variant of the tasks::TaskOp type, by
removing the optional task ID we kept around as part of the "before"
member. This task ID is not necessary, because the task ID does not
change as part of the update, so we might as well work with that of the
"update" attribute.