Rename ui::dialog::Data::task
This patch renames the task member of the dialog module's Data type from
the generic term 'task' to the more specific 'to_edit'. This rename both
makes it clearer what this task is for while also preventing naming
conflicts in the future, when we intend replace the adjacent 'task_id'
task ID with an Rc<Task> called 'task'.