Use Rc<Task> in more places
This change pushes the usage of Rc<Task> over task IDs into more places.
Most importantly, we adjust most of the interfaces of the Tasks type to
work with Rc<Task> objects. That in turn has trickle down effects on
many other parts of the program. We keep task IDs around for the time
being where that is necessary, but remove them where not.