Switch more clients over to using Db::find_item()
Now that the Db's push() and insert() methods return EntryMut objects,
we can use that to our advantage and replace more Db::find() calls,
which require an ID instead of an actual object reference, with
Db::find_item(). Doing so moves us one step closer to getting rid of
Db::find() altogether, which is a prerequisite of eliminating numerical
IDs from major parts of the code base.