Re-introduce ID member to Task type
A while back we removed the ID member from the Task type, basically
arguing that IDs should be "externally" managed, in our case by the
database introduced alongside.
The program has changed in design and requirements, and we presumably
have a less opinionated view on how parts should interact. So after some
reflection, it has become our believe that having IDs embedded in the
object in question will result in the cleanest design.
We never adjusted our tags to have externally managed IDs (although we
tried) and so the only type that has an associated but not embedded ID
is the Task type. Fixing this perceived design flaw, this change
re-introduces an ID member to it.
Having an ID in there raises the obvious question of who ensures
uniqueness. Long term, the goal is for us to use excessively large (and
supposedly "globally" unique) UUIDs. However, until we get there this
change embeds a quick and dirty (and temporary!) bump allocator to the
Task type itself. Conceptually, that cannot guarantee uniqueness in the
presence of persisted IDs moving forward, but we split the available ID
space in half and only use the upper half for on-demand allocation.