Remove ID logic from Db
So far our Db types has taken care of associating an ID with each item
it contains. These IDs served as references of sorts, that could later
be used to look up items in the database.
With recent changes we have started to move away from relying on IDs and
instead used "true" references, in the form of Rc objects that can
easily be cloned. By now most parts of the program have migrated over to
using Rc instances directly a so with this change we remove all ID
related logic from the Db type.
Note that as a consequence of the ID removal, the Entry and EntryMut
types effectively have no use at the moment and could be replaced with
"direct" references. However, we anticipate future use cases and want to
minimize unnecessary churn for the time being and so we keep them
around for now.