Remove AllocId impl for BTreeSet
A recent change removed the ID tracking from our Db type. This step also
eliminated the need for said type to allocate and free IDs. Given that
it was the only entity using BTreeSet based allocation, no one is using
the AllocId impl for the BTreeSet type anymore and so this change
removes it.