Adjust Db to work with Rc objects
So far our Db type stored arbitrary data of any type T. While we keep it
working for arbitrary T, the database now stores each item in an Rc and
treats it as having an "identity". On top of that, it enforces that only
one copy of any Rc can exist within its purview.
Given that Task objects were already stored within an Rc most of the
time, we effectively just move knowledge of that fact around: earlier it
was the Tasks layer that effectively managed those instances, while now
we move this responsibility into the Db type.
One consequence of this move is that the Db type's API has to be
adjusted: it no longer support mutating of elements directly, via
methods such as Db::get_mut. Furthermore, in addition to infallible
methods for item insertion (insert and push), we introduce fallible
ones. The infallible variants make sure to perform heap allocation
(i.e., wrapping of items in an Rc) themselves, while the fallible ones
allow for insertion of Rc objects.
Another one is that we can get rid of the Cmp trait and CmpRc type,
because a Db now has knowledge about equality of its items (Rc objects)
baked in.