Make tasks::Id private
With recent changes we have removed traces of task IDs throughout the
code base, replacing them with Rc<Task> instances. This change marks the
logical conclusion of this transformation, marking the respective type
as private to the containing module, ensuring that usage of it will not
proliferate again outside of it.
Note, however, that for one test in particular, we still have an
accessor that exposes a task's ID object across module boundaries. We
want to keep that, but duly note that it is a test-only constellation.