Remove logic for handling "regular" task IDs
The change switching us to using UUIDs for task identification by
default specifically included logic to be able to deal with "regular"
IDs, converting them to UUIDs during program load, in order to allow for
a graceful migration.
With this change we assume that migration has happened and remove this
logic. Moving forward it will only be possible to load tasks using UUIDs
as identifier.