Use anyhow for error handling
So far we use (or misused) std::io::Error for our error reporting
purposes. While there is basic error chaining support, that only allows
attachment of a few bits of information. As a result, errors reported by
the program are terse and without context, making it incredibly hard to
root cause issues.
This change switches the program over to using anyhow for error
management. The crate allows easy contextualization of errors at
different layers and presents them in an easy to understand format.
Along with the transition from using one error type to another, we also
add additional error context where it is beneficial.