Store ID in ser::Task
This change adjusts our serializable and deserializable Task object to
contain a task ID. So far we have shied away from doing so, because a
lot of tests compare such objects for equality to check functionality,
and having an ID in it complicates matters unnecessarily.
Right now we work around this problem by excluding this newly added
field from any equality checks. This works fine, but it is not the
cleanest solution. However, it is the better design and we need the ID
to be included there for upcoming changes.