Introduce backend abstraction
Right now we store all our data in JSON encoded form. We want to make
true of the promise of enabling CalDAV support in a first version, and
as such we would like to serialize tasks and meta data to iCalendar
files instead.
To make that possible, this change introduces a new abstraction to the
program, a Backend. It is a trait that we intend to use for
serialization and deserialization matters in the existing state
management logic.