Add iCal support for Task objects
This change adds iCal serialization and deserialization support for Task
objects. We piggy back on icalendar types, Calendar and Todo
specifically, from/into which we convert. Objects of types are what we
actually persist, but then store certain task attributes in properties
on them.
The reason for doing this chain of conversions is that CalDav seems to
work with Calendar objects. Those are only containers for components.
One such component type is a Todo. Our internally used Task type
effectively maps to such a Todo.