Add iCal support for TasksMeta objects
Along with Task objects themselves, the second task related type we need
to have support for in the iCal backend is TasksMeta. Objects of this
type persist task related meta data and this data has to be kept in sync
with the tasks themselves.
This change adds support for serializing and deserializing TasksMeta
objects using the iCal backend. As before, we have to store this meta
data in a Todo, which in turn is wrapped in a Calendar. We tried adding
the necessary properties to the Calendar object directly, but they are
not honored correctly when deserializing. While this may be a bug in the
icalendar crate, pretty much either approach is fine with us and could
conceivably be changed later on if a need arises.