Use iCal backend for saving and loading tasks and task meta data
With recent changes we have created a fully functional backend for
storing tasks and the associated meta data in iCalendar format, instead
of using JSON.
With this change we make this backend the one we use for saving task
state, and make it the default one when loading as well. We still keep
the option of loading data from JSON (if the iCal load fails) in order
to facilitate live conversions from the program, but this feature will
be removed subsequently.