Remove support for saving or loading tasks in JSON format
When we rolled out the support for creating ics files instead of JSON,
we still kept the option of loading JSON around as a temporary fall
back.
With this change we remove that fall back and rely solely on the
iCalendar backend for all storage matters for tasks. The UI, of course,
is still using JSON.