Constrain serde_json usage to Json backend
A bunch of our tests still "hardcode" the usage of serde_json
serialization and deserialization functions, when really what we are
using in the program is the indirection through the Json backend. This
change converts those tests over to using the backend as well, to make
them as representative as they can be again, while also constraining the
usage of serde_json inside the Json backend itself.